+++ cmd/hidden_lake/messenger/_daemon/install_hlm.sh
@@ -1,0 +1,25 @@
+# root mode
+
+echo "
+[Unit]
+Description=HiddenLakeMessenger
+
+[Service]
+ExecStart=/root/hlm_amd64_linux -path=/root
+Restart=always
+RestartSec=10
+
+[Install]
+WantedBy=multi-user.target
+" > /etc/systemd/system/hidden_lake_messenger.service
+
+cd /root && \
+    rm -f hlm_amd64_linux && \
+    wget https://github.com/number571/go-peer/releases/download/v1.5.18/hlm_amd64_linux && \
+    chmod +x hlm_amd64_linux
+
+chmod +x /root/hls_amd64_linux
+systemctl daemon-reload
+systemctl enable hidden_lake_messenger.service
+systemctl restart hidden_lake_messenger.service
+watch -c SYSTEMD_COLORS=1 systemctl status hidden_lake_messenger.service