+++ cmd/hidden_lake/messenger/_daemon/stop_hlm.sh
@@ -1,0 +1,3 @@
+# root mode
+
+systemctl stop hidden_lake_messenger.service